FCA_Task/requirements.txt
asn1crypto==1.5.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==2.2.1
cryptography==41.0.7
et-xmlfile==1.1.0
filelock==3.13.1
idna==3.6
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
platformdirs==3.11.0
pyarrow==15.0.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
snowflake-connector-python==3.7.0
snowflake-snowpark-python==1.12.1
sortedcontainers==2.4.0
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
